Fixes tests from PR feedback
